media: fpga-read: support for 1920x1080x60 & 1280x1024x60 resolution
Addition of the 2 supported modes for faster data communication
speeds. As the resolution goes high phy frequency increases.

In 1920x1080x60, phy frequency of 1100 MHz is required.
In 1280x1024x60, phy frequency of 700 MHz is required.

Both are running on the 4 Data lanes. 1280x1024x60 can be run
on 2 Data lanes, too.

Data lanes can be updated from the dtbs.